Add files via upload
Though not all of these scripts are needed for you to achieve the same effect for one way platforms, these are the scripts I have on my 2D Character that allow for the Character to go through one way platforms, hang from the ledge, be able to jump down automatically through a one way platform if downward jumping, be able to fall through a one way platform if standing on the platform and not downward jumping, and be able to move horizontally while jumping through a one way platform.

If you aren't using all my scripts and have your own Character type script that all scripts for the player would inherit from, it is recommended that if you copy the CollisionCheck() method and paste in your Character script as both the Jump script and HorizontalMovement scripts will use this method to handle the logic to cancel the downward jump as well as move horizontally while jumping through a one way platform. 

If you are not familiar with inheritance (though it is highly recommended you become somewhat familiar with it) and have your horizontal movement and jump scripts both inherit from Monobehaviour, then you can simply copy that method from my Character script and paste it in both your jump script and horizontal movement scripts so you can have the features of being able to jump downwards automatically through a one way platform as discussed in the video as well as move horizontally through them.

Of course if you don't need or want those features like Dead Cells has them, then you can skip all that together so you can perform the basic logic needed to go through one way platforms and ledge hanging capabilities.